disabled wittness generation, work without public input
