+++ korrekt/src/circuit_analyzer/mod.rs
@@ -1,4 +1,3 @@
 pub mod abstract_expr;
 pub mod analyzer;
-//pub mod layouter;
-//pub mod shape;
+pub mod analyzable;
