+++ korrekt/src/test/mod.rs
@@ -1,3 +1,2 @@
 pub mod integration_tests;
-pub mod integration_tests_mockprover_based;
 pub mod pse_test_row_diff_is_zero;
