--- rust-toolchain
@@ -1,1 +1,0 @@
-nightly-2022-08-23
